+++ src/commands/expressions/operands/value.ts
@@ -1,5 +1,4 @@
 import type { AttributeValue } from "../../../types.js";
-import type { AttributeNames } from "../../attributes/names.js";
 import type { AttributeValues } from "../../attributes/values.js";
 import type { IOperand } from "./type.js";
 
@@ -19,10 +18,8 @@ export class ValueOperand<T extends AttributeValue = AttributeValue>
     this.value = value;
   }
 
-  substitute(params: {
-    names: AttributeNames;
-    values: AttributeValues;
-  }): string {
+  // TODO: make the names optional
+  substitute(params: { values: AttributeValues }): string {
     const { values } = params;
     return values.substitute(this.value);
   }